Bluetooth: controller: ULL: Make adv_sync_hdr_set_clear able to update PDU in place
Enabling or disabling transmission of Constant Tone Extension with
periodic advertising PDUs requires update of a PDU content.
CTE_INFO field of periodic advertising PDU must be filled with
appropriate information. That operation may be done when
periodic advertising PDU (or chain of PDUs) is filled with other
data and wasn't scheduled for transmission. To be memory efficient
PDUs not yet scheduled for transmission are updated in place.

Function adv_sync_hdr_set_clear was implemented that it
updated content of extended header flags directly in PDU.
Now flags are updated at the end of the function when other
parts of PDU are already there.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>